requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.10
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
validators==0.34.0
